\name{feedquery}
\alias{feedquery}
\title{Buildup string for feedquery.}
\usage{feedquery(url, params)
}
\description{Buildup string for feedquery. Function has partly been taken from getForm function in
\code{XML} package (Duncan Temple Lang)}
\author{Mario Annau}
\seealso{\code{\link{xmlNode}}}
\arguments{\item{url}{character}
\item{params}{list which contains feed parameters}
}

