\name{getEmpty}
\alias{getEmpty}
\alias{getEmpty.WebCorpus}
\title{Get Empty Corpus Elements
Retrieve content of all empty Corpus elements}
\usage{
  getEmpty(x, ...)
}
\arguments{
  \item{x}{object of type \code{\link{WebCorpus}}}

  \item{...}{additional parameters to PostFUN}
}
\description{
  Get Empty Corpus Elements Retrieve content of all empty
  Corpus elements
}

