\name{getMainText}
\alias{getMainText}
\title{Get Main Text from Annotated HTML Tree
Main Text is obtained from Tree -Subnode where threshold > threshold and
textlength is at maximum}
\usage{
  getMainText(xml, threshold)
}
\arguments{
  \item{xml}{object of class xmlNode}

  \item{threshold}{minimum threshold needed to be
  considered}
}
\description{
  Get Main Text from Annotated HTML Tree Main Text is
  obtained from Tree -Subnode where threshold > threshold
  and textlength is at maximum
}
\author{
  Mario Annau
}
\seealso{
  \code{\link{extractContentDOM}}, \code{\link{xmlNode}}
}

