\name{getURL}
\alias{getURL}
\title{Copy of RCurl:::getURL() including a little bugfix for the .encoding parameter.}
\usage{
getURL(url, ..., .opts = list(), write = basicTextGatherer(.mapUnicode =
  .mapUnicode), curl = getCurlHandle(), async = length(url) > 1,
  .encoding = integer(), .mapUnicode = TRUE)
}
\arguments{
  \item{url}{see \code{\link[RCurl]{getURL}}}

  \item{.opts}{see \code{\link[RCurl]{getURL}}}

  \item{write}{see \code{\link[RCurl]{getURL}}}

  \item{curl}{see \code{\link[RCurl]{getURL}}}

  \item{async}{see \code{\link[RCurl]{getURL}}}

  \item{.encoding}{see \code{\link[RCurl]{getURL}}}

  \item{.mapUnicode}{see \code{\link[RCurl]{getURL}}}

  \item{...}{see \code{\link[RCurl]{getURL}}}
}
\description{
For the full description of getURL refer to
\code{\link[RCurl]{getURL}}.
}
\seealso{
\code{\link[RCurl]{getURL}}
}

