% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/source.R
\name{LiberationSource}
\alias{LiberationSource}
\alias{readLiberationSource}
\title{Get news data from french Liberation News Paper (\url{http://rss.liberation.fr/rss}).}
\usage{
LiberationSource(query = "latest", ...)
}
\arguments{
\item{query}{feed to be retrieved, defaults to 'latest'}

\item{...}{additional parameters to \code{\link{WebSource}}}
}
\value{
WebXMLSource
}
\description{
Get news data from french Liberation News Paper (\url{http://rss.liberation.fr/rss}).
}
\examples{
\dontrun{
corpus <- WebCorpus(LiberationSource("latest"))
}
}
\author{
Mario Annau
}
\seealso{
\code{\link{WebSource}}
}

