% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tm.plugin.webmining-package.R
\docType{data}
\name{nytimes_appid}
\alias{nytimes_appid}
\title{AppID for the NYtimes-API.}
\description{
USED ONLY FOR PACKAGE TESTING. PLEASE DOWNLOAD YOUR OWN KEY AT \url{http://developer.nytimes.com/}!!!
}
\author{
Mario Annau
}
\keyword{data}

