\name{removeTags}
\alias{removeTags}
\title{Remove specified tags from (XML) Document Tree.}
\usage{removeTags(xmldoc, tags)
}
\description{Remove specified tags from (XML) Document Tree.
Tags and all of its inner content will be removed.}
\author{Mario Annau}
\seealso{\code{\link{extractContentDOM}}}
\arguments{\item{xmldoc}{xmlDoc object of class xmlDoc}
\item{tags}{character vector which specifies tags to remove}
}

