\name{source.update}
\alias{source.update}
\alias{source.update.WebHTMLSource}
\alias{source.update.WebJSONSource}
\alias{source.update.WebXMLSource}
\title{Update WebXMLSource/WebHTMLSource/WebJSONSource
Typically, update is called from \code{link{corpus.update}} and refreshes \code{$Content} in
Source object.}
\usage{
  source.update(x)
}
\arguments{
  \item{x}{Source object to be updated}
}
\description{
  Update WebXMLSource/WebHTMLSource/WebJSONSource
  Typically, update is called from
  \code{link{corpus.update}} and refreshes \code{$Content}
  in Source object.
}

