% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trimWhiteSpaces.R
\name{trimWhiteSpaces}
\alias{trimWhiteSpaces}
\title{Trim White Spaces from Text Document.}
\usage{
trimWhiteSpaces(txt)
}
\arguments{
\item{txt}{character}
}
\description{
Transformation function, actually equal to stripWhiteSpace
applicable for simple strings using Perl parser
}
\author{
Mario Annau
}
\seealso{
\code{\link{stripWhitespace}}
}

