% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/source.R
\name{YahooNewsSource}
\alias{YahooNewsSource}
\alias{readYahooHTML}
\title{Get news data from Yahoo! News (\url{https://news.search.yahoo.com/search/}).}
\usage{
YahooNewsSource(query, params = list(p = query), ...)
}
\arguments{
\item{query}{words to be searched in Yahoo News, multiple words must be separated by '+'}

\item{params,}{additional query parameters, see \url{http://developer.yahoo.com/rss/}}

\item{...}{additional parameters to \code{\link{WebSource}}}
}
\value{
WebXMLSource
}
\description{
Currently, only a maximum of 10 items can be retrieved.
}
\examples{
\dontrun{
corpus <- Corpus(YahooNewsSource("Microsoft"))
}
}
\author{
Mario Annau
}
\seealso{
\code{\link{WebSource}}
}

