\name{corpus.update}
\alias{corpus.update}
\alias{corpus.update.WebCorpus}
\title{Update Corpus
Generic function to update a text \code{\link{Corpus}}}
\usage{
  corpus.update(x, ...)
}
\arguments{
  \item{x}{object of type \code{\link{WebCorpus}}}

  \item{...}{additional parameters to \code{\link{Corpus}}
  function}
}
\description{
  Update Corpus Generic function to update a text
  \code{\link{Corpus}}
}

