\name{extract}
\alias{extract}
\alias{extract.PlainTextDocument}
\title{Extract Main Content from Text Documents}
\usage{
  extract(x, extractor, ...)
}
\arguments{
  \item{x}{PlainTextDocument}

  \item{extractor}{default extraction function to be used,
  defaults to \code{\link{extractContentDOM}}}

  \item{...}{additional parameters to extractor function}
}
\description{
  Extract Main Content from Text Documents
}

