\name{trimWhiteSpaces}
\alias{trimWhiteSpaces}
\title{Trim White Spaces from Text Document.}
\usage{trimWhiteSpaces(txt)
}
\description{Trim White Spaces from Text Document.
Transformation function, actually equal to stripWhiteSpace 
applicable for simple strings using Perl parser}
\author{Mario Annau}
\seealso{\code{\link{stripWhitespace}}}
\arguments{\item{txt}{character}
}

