\name{trimWhiteSpaces}
\alias{trimWhiteSpaces}
\title{Trim White Spaces from Text Document.
Transformation function, actually equal to stripWhiteSpace
applicable for simple strings using Perl parser}
\usage{
  trimWhiteSpaces(txt)
}
\arguments{
  \item{txt}{character}
}
\description{
  Trim White Spaces from Text Document. Transformation
  function, actually equal to stripWhiteSpace applicable
  for simple strings using Perl parser
}
\author{
  Mario Annau
}
\seealso{
  \code{\link{stripWhitespace}}
}

