\docType{data}
\name{yahoonews}
\alias{yahoonews}
\title{WebCorpus retrieved from Yahoo! News for the search term "Microsoft"
through the YahooNewsSource. Length of retrieved corpus is 20.}
\description{
WebCorpus retrieved from Yahoo! News for the search term
"Microsoft" through the YahooNewsSource. Length of
retrieved corpus is 20.
}
\examples{
#Data set has been generated as follows:
\dontrun{
yahoonews <- WebCorpus(YahooNewsSource("Microsoft"))
}
}
\author{
Mario Annau
}
\keyword{data}

