\docType{data}
\name{yahoonews}
\alias{yahoonews}
\title{This is data to be included in my package}
\description{
  This is data to be included in my package
}
\examples{
#Data set has been generated as follows:
\dontrun{
yahoonews <- WebCorpus(YahooNewsSource("Microsoft"))
}
}
\author{
  Mario Annau
}
\references{
  \url{search.twitter.com}
}
\keyword{data}
\keyword{datasets}

